+++ robotic_skin/calibration/error_functions.py
@@ -1,217 +1,8 @@
-# import logging
 import numpy as np
-import robotic_skin.const as C
 import pyquaternion as pyqt
+import robotic_skin.const as C
 from robotic_skin.calibration.utils.quaternion import np_to_pyqt
-# from robotic_skin.calibration.utils import n2s
-
-
-def estimate_acceleration(kinematic_chain, i_rotate_joint, i_su,
-                          joint_angular_velocity, joint_angular_acceleration=0,
-                          max_angular_velocity=0, current_time=0, method='analytical'):
-    r"""
-    Compute an acceleration value from positions.
-    .. math:: `a = \frac{f({\Delta t}) + f({\Delta t) - 2 f(0)}{h^2}`
-
-    This equation came from Taylor Expansion to get the second derivative from f(t).
-    .. math:: f(t+{\Delta t}) = f(t) + hf^{\prime}(t) + \frac{h^2}{2}f^{\prime\prime}(t)
-    .. math:: f(t-{\Delta t}) = f(t) - hf^{\prime}(t) + \frac{h^2}{2}f^{\prime\prime}(t)
-
-    Add both equations and plug t=0 to get the above equation
-
-    Arguments
-    ------------
-    `kinematic_chain`: `robotic_skin.calibration.kinematic_chain.KinematicChain`
-        Robot's Kinematic Chain
-    `i_rotate_joint`: `int`
-        dof `d`
-    `i_su`: `int`
-        `i`th SU
-    `joint_angular_velocity`: `float`
-        Angular velocity
-    'max_angular_velocity': 'float'
-        Maximum angular velocity
-    `current_time`: `float`
-        Current Time
-    `method`: `str`
-        Determines if we are using `analytical`, `normal_mittendorfer` or `mittendorfer`
-        methods (which we modified due to some possible missing terms).
-    """
-    methods = ['analytical', 'mittendorfer', 'normal_mittendorfer']
-    if method not in methods:
-        raise ValueError(f'There is no method called {method}\n' +
-                         f'Please Choose from {methods}')
-
-    rs_T_su = kinematic_chain.compute_su_TM(
-        i_su=i_su, pose_type='current')
-
-    dof_T_su = kinematic_chain.compute_su_TM(
-        start_joint=i_rotate_joint,
-        i_su=i_su,
-        pose_type='current')
-
-    dof_w_su = np.array([0, 0, joint_angular_velocity])
-    dof_alpha_su = np.array([0, 0, joint_angular_acceleration])
-    su_g, su_Ac, su_At = compute_acceleration_analytically(
-        inert_w_body=dof_w_su,
-        inert_r_body=dof_T_su.position,
-        inert_alpha_body=dof_alpha_su,
-        body_R_inert=dof_T_su.R.T,
-        body_R_world=rs_T_su.R.T,
-        coordinate='body')
-
-    if method == 'analytical':
-        return su_g + su_Ac + su_At
-
-    # The following will run if method is mittendorfer's method
-    su_At = compute_tangential_acceleration_numerically(
-        kinematic_chain=kinematic_chain,
-        i_rotate_joint=i_rotate_joint,
-        i_su=i_su,
-        joint_angular_velocity=joint_angular_velocity,
-        max_angular_velocity=max_angular_velocity,
-        current_time=current_time)
-
-    if method == 'normal_mittendorfer':
-        # return su_At
-        return su_g + su_At
-
-    # Every joint rotates along its own z axis, one joint moves at a time
-    return su_g + su_Ac + su_At
-
-
-def centripetal_acceleration(r, w):
-    r"""
-    .. math:: `a = \omega \times \omega \times r`
-
-    Arguments
-    ------------
-    `r`: `float`
-        Position vector r of the body in the inertial coordinate
-    `w`: `float`
-        Angular Velocity of an axis in the inertial coordinate.
-    """
-    # a = w x w x r
-    return np.cross(w, np.cross(w, r))
-
-
-def tangential_acceleration(r, alpha):
-    """
-    .. math:: `a = \alpha \times r`
-
-    Arguments
-    ----------
-    `r`: `float`
-        Position vector r of the body in the inertial coordinate
-    `alpha`: `float`
-        Angular Acceleration of an axis in the inertial coordinate.
-    """
-    # a = al x r
-    return np.cross(alpha, r)
-
-
-def compute_acceleration_analytically(inert_w_body, inert_r_body, inert_alpha_body,
-                                      body_R_inert=None, body_R_world=None, inert_R_world=None,
-                                      coordinate='body'):
-    """
-    There are 3 coordinates to remember.
-    1. World Frame (Fixed to the world)
-    2. Inertial Frame (each Rotating Joint Coordinate)
-    3. Body Frame (SU Coordinate)
-
-    We use the following notation `coord1_variable_coord2`.
-    This represents the variable of coord2 defined in coord1.
-
-    For example, SU's linear accelerations can be represented as rs_a_su,
-    which means acceleration a measured in SU coordinate defined in the world coordinate.
-
-    We use the same notation for the rotation matrix `coord1_R_coord2`,
-    but this represents that it rotations some variables from `coord2` to `coord1`.
-
-    For example, if you want the gravity in the su coordinate,
-    su_g = su_R_rs * rs_g
-    """
-    world_g = np.array([0, 0, 9.80])
-    inert_Ac_body = centripetal_acceleration(r=inert_r_body, w=inert_w_body)
-    inert_At_body = tangential_acceleration(r=inert_r_body, alpha=inert_alpha_body)
-
-    if coordinate == 'body':
-        if body_R_inert is None or body_R_world is None:
-            raise ValueError('You must provide Rotation matrices body_R_inert and body_R_world')
-        # Convert to body coordinate
-        body_g = np.dot(body_R_world, world_g)
-        body_Ac = np.dot(body_R_inert, inert_Ac_body)
-        body_At = np.dot(body_R_inert, inert_At_body)
-        return body_g, body_Ac, body_At
-
-    elif coordinate == 'world':
-        if inert_R_world is None:
-            raise ValueError('You must provide a Rotation Matrix inert_R_world')
-        # Convert to world coordinate
-        world_Ac = np.dot(inert_R_world.T, inert_Ac_body)
-        world_At = np.dot(inert_R_world.T, inert_At_body)
-        return world_g, world_Ac, world_At
-
-    else:
-        raise KeyError(f'Coordinate name "{coordinate}" is invalid\n' +
-                       'Please choose from "body", "inertial", or "world"')
-
-
-def compute_2nd_order_derivative(x_func, t=0, dt=0.001):
-    # dt should be small value, recommended to use 1/(1000 * freq)
-    return ((x_func(t+dt) + x_func(t-dt) - 2*x_func(t)) / (dt**2))
-
-
-def compute_tangential_acceleration_numerically(kinematic_chain, i_rotate_joint, i_su,
-                                                joint_angular_velocity, max_angular_velocity,
-                                                current_time):
-    """
-    Returns tangential acceleration in RS coordinate.
-    The acceleration is computed by taking 2nd derivative of the position.
-    This small change in position in Rotating Coordinate is only in the
-    tangential direction. Thus, you can only compute the tangential acceleration,
-    from this method.
-
-
-    Arguments
-    ---------
-    `kinematic_chain`: `robotic_skin.calibration.kinematic_chain.KinematicChain`
-        Robot's Kinematic Chain
-    'i_rotate_joint': 'int'
-        dof 'd'
-    `i`: `int`
-        imu `i`
-    'max_angular_velocity': 'float'
-        Maximum angular velocity
-    `current_time`: float`
-        Current Time
-    """
-    def su_position_in_sinmotion(t):
-        angle = (max_angular_velocity / (2*np.pi*C.PATTERN_FREQ)) * (1 - np.cos(2*np.pi*C.PATTERN_FREQ * t))
-        kinematic_chain.init_temp_TM(i_joint=i_rotate_joint, additional_pose=angle)
-        T = kinematic_chain.compute_su_TM(i_su=i_su, pose_type='temp')
-        return T.position
-
-    # Compute the Acceleration from 3 close positions
-    rs_A = compute_2nd_order_derivative(x_func=su_position_in_sinmotion, t=current_time)
-    # Compute current Angle during sinuosoidal motion
-    angle = (max_angular_velocity / (2*np.pi*C.PATTERN_FREQ)) * (1 - np.cos(2*np.pi*C.PATTERN_FREQ * current_time))
-    # Get current transformation matrix of SU defined in dof coordinate
-    kinematic_chain.init_temp_TM(i_joint=i_rotate_joint, additional_pose=angle)
-    dof_T_su = kinematic_chain.compute_su_TM(i_su=i_su, pose_type='temp', start_joint=i_rotate_joint)
-    rs_T_dof = kinematic_chain.compute_joint_TM(i_joint=i_rotate_joint, pose_type='temp')
-    dof_A = np.dot(rs_T_dof.R.T, rs_A)
-    # Tangential Vector (to the circular motion)
-    e_t = np.cross([0, 0, 1], dof_T_su.position)
-    e_t = e_t / np.linalg.norm(e_t)
-
-    # Only retrieve the tangential element of dof_A,
-    # because dof_A also includes unnecessary centripetal acceleration
-    dof_At = e_t * np.dot(e_t, dof_A)
-    su_At = np.dot(dof_T_su.R.T, dof_At)
-
-    # plot_projected_acceleration(dof_A, e_r, e_t, dof_At, dof_Ac)
-    return su_At
+from robotic_skin.calibration.utils.rotational_acceleration import estimate_acceleration
 
 
 class ErrorFunction():
@@ -420,7 +211,7 @@ class MaxAccelerationErrorFunction(ErrorFunction):
                 joints = data[:, 3:10]
                 times = data[:, 10]
                 joint_angular_accelerations = data[:, 11]
-                max_angular_velocity = data[0, 12]
+                # max_angular_velocity = data[0, 12]
                 joint_angular_velocities = data[:, 13]
 
                 n_eval = 4
@@ -445,8 +236,8 @@ class MaxAccelerationErrorFunction(ErrorFunction):
                         i_su=i_su,
                         joint_angular_velocity=joint_angular_velocity,
                         joint_angular_acceleration=joint_angular_acceleration,
-                        max_angular_velocity=max_angular_velocity,
                         current_time=time,
+                        angle_func=self.angle_func,
                         method=self.method)
 
                     # logging.debug(f'[{pose}, {joint}, {su}@Joint{i_joint}]\t' +
@@ -456,6 +247,18 @@ class MaxAccelerationErrorFunction(ErrorFunction):
                     n_data += 1
 
         return e2/n_data
+
+    def angle_func(self, t: int):
+        """
+        Computes current joint angle at time t
+        joint is rotated in a sinusoidal motion during MaxAcceleration Data Collection.
+
+        Parameters
+        ------------
+        `t`: `int`
+            Current time t
+        """
+        return (C.MAX_ANGULAR_VELOCITY / (2*np.pi*C.PATTERN_FREQ)) * (1 - np.cos(2*np.pi*C.PATTERN_FREQ * t))
 
 
 class CombinedErrorFunction(ErrorFunction):
