+++ robotic_skin/examples/calibration/plot_data.py
@@ -15,50 +15,85 @@ CONFIGDIR = os.path.join(REPODIR, 'config')
 
 
 def is_first(i):
+    """
+    Returns whether it's the first row/column
+    """
     return i == 0
 
 
 def is_last(i, n):
+    """
+    Returns whether it's the last row/column
+    """
     return i == (n - 1)
 
 
 def plot_side_by_side(y1: np.ndarray, y2: np.ndarray,
-                      xlabel: str, title1: str, title2: str, show=True):
+                      title1: str, title2: str, xlabel: str,
+                      ylabels: List[str] = ['ax', 'ay', 'az'],
+                      show=True):
+    """
+    Plot data side by side: `y1` on the left and `y2` on the right
+
+    Arguments
+    ----------
+    `y1`: `np.ndarray`
+        Data. This function mainly targets time series data
+        Shape = (length, data)
+    `y2`: `np.ndarray`
+        Data. This function mainly targets time series data
+        Shape = (length, data)
+    `title1`: `str`
+        Title of y1
+    `title2`: `str`
+        Title of y2
+    `xlabel`: `str`
+        Normally, t's Time [s] or No. Data Points
+    `ylabels`: `List[str]`
+        Data's label
+    `show`: `bool`
+
+    """
     if y1.size == 0 or y2.size == 0:
-        print('Data cannot be empty')
+        print('y1 and y2 cannot be empty')
         return
 
     if y1.shape != y2.shape:
-        raise ValueError('Arguments must be same size')
+        raise ValueError('y1 and y2 must be same size')
 
     n_data = y1.shape[0]
     n_row = y1.shape[1]
     x = np.arange(n_data)
 
-    ylabels = ['ax', 'ay', 'az'] + ['w', 'alpha']
+    if len(ylabels) != n_row:
+        raise ValueError(f'num of ylabels must be {n_row}')
+
     fig = plt.figure(figsize=(10, 8))
 
+    # For All Data/Rows
     for i_row in range(n_row):
         ax_left = fig.add_subplot(n_row, 2, 2*i_row+1)
         ax_right = fig.add_subplot(n_row, 2, 2*i_row+2)
-
+        # Compute y_min and y_max for the combined data y
         y = np.hstack((y1[:, i_row], y2[:, i_row]))
         y_min = np.min(y)
         y_max = np.max(y)
         y_range = y_max - y_min
         y_min -= 0.1 * y_range
         y_max += 0.1 * y_range
-
+        # Plot Left
         ax_left.plot(x, y1[:, i_row])
-        ax_left.set_ylabel(ylabels[i_row])
         ax_left.set_ylim([y_min, y_max])
-
+        ax_left.set_ylabel(ylabels[i_row])
+        # Plot Right
         ax_right.plot(x, y2[:, i_row])
         ax_right.set_ylim([y_min, y_max])
 
+        # Set Title at the top
         if is_first(i_row):
             ax_left.set_title(title1)
             ax_right.set_title(title2)
+        # Set Xlabel at the bottom
         if is_last(i_row, n_row):
             ax_left.set_xlabel(xlabel)
             ax_right.set_xlabel(xlabel)
@@ -67,9 +102,35 @@ def plot_side_by_side(y1: np.ndarray, y2: np.ndarray,
         plt.show()
 
 
-def plot_methods_at_once(y_dict: dict, ylabels: List[str], xlabel: str,  # noqa:C901
-                         title: str, x: np.ndarray = None,
-                         show=True, save=False):
+def plot_in_one_graph(y_dict: dict, ylabels: List[str], xlabel: str,  # noqa:C901
+                      title: str, x: np.ndarray = None,
+                      show=True, save=False):
+    """
+    Plot all y_dict data in 1 graph.
+    Each y_dict data is assumed to be 2 dimension.
+    Shape=(length, data)
+
+    Arguments
+    ----------
+    `x`: `np.ndarray`
+        x axis data
+    `xlabel`:
+        Normally it's Time [s] or No. Data Points
+    `y_dict`: `dict`
+        Data stored in a dictionary.
+        Keys are the names of the data. (Ex. Method names)
+        Values include the actual data.
+        This function mainly targets time series data
+        y_dict['key'].shape = (length, data)
+    `ylabels`: `List[str]`
+        Labels of the data
+    `title`: `str`
+        Title
+    `show`: `bool`
+        Show plot
+    `save`: `bool`
+        Save plot
+    """
     if not isinstance(y_dict, dict):
         raise ValueError('y_dict be a dictionary')
     if len(y_dict) == 0:
@@ -99,33 +160,35 @@ def plot_methods_at_once(y_dict: dict, ylabels: List[str], xlabel: str,  # noqa:
 
     for i_row in range(n_row):
         ax = fig.add_subplot(n_row, 1, i_row+1)
-
+        # Plot and combine all data at the same time
         y = np.array([])
         for method, data in y_dict.items():
             ax.plot(x, data[:, i_row], label=method)
             y = np.hstack((y, data[:, i_row]))
 
+        # Compute y_min and y_max for the combined data y
         y_min = np.min(y)
         y_max = np.max(y)
         y_range = y_max - y_min
         y_min -= 0.1 * y_range
         y_max += 0.1 * y_range
-
+        # For all Row
         ax.set_ylabel(ylabels[i_row])
         ax.set_ylim([y_min, y_max])
-
+        # Set title at the top
         if is_first(i_row):
             ax.set_title(title)
             ax.legend(loc='best')
-
+        # Set xlable at the bottom
         if is_last(i_row, n_row):
             ax.set_xlabel(xlabel)
 
     if show:
         plt.show()
 
+    # Save to IMGDIR/plot_in_one_graph directory
     if save:
-        dirname = os.path.join(IMGDIR, 'plot_methods_at_once')
+        dirname = os.path.join(IMGDIR, 'plot_in_one_graph')
         if not os.path.exists(dirname):
             os.makedirs(dirname)
         savepath = os.path.join(dirname, title + '.png')
@@ -143,9 +206,8 @@ def verify_if_noise_is_added_correctly(args):
     # Add Noise
     sigma = 1.0
     outlier_ratio = 0.5
-    utils.add_outlier(data_noise, 'static', sigma=sigma, outlier_ratio=outlier_ratio)
-    utils.add_outlier(data_noise, 'dynamic', sigma=sigma, outlier_ratio=outlier_ratio)
-    utils.add_outlier(data_noise, 'constant', sigma=sigma, outlier_ratio=outlier_ratio)
+    data_types = ['static', 'dynamic', 'constant']
+    utils.add_outlier(data_noise, data_types, sigma=sigma, outlier_ratio=outlier_ratio)
 
     pose_names = list(data.constant.keys())
     joint_names = list(data.constant[pose_names[0]].keys())
@@ -218,32 +280,35 @@ def verify_if_noise_is_added_correctly(args):
 
 
 def verify_estimated_accelerations_for_dynamic_datacollection(args):
+    # Preparation for the initialization of KinematicChain
     robot_configs = utils.load_robot_configs(args.configdir, args.robot)
-
     datadir = utils.parse_datadir(args.datadir)
     data, imu_mappings = utils.load_data(args.robot, datadir)
-
+    # Initialize KinematicChain
     kinematic_chain = construct_kinematic_chain(
         robot_configs=robot_configs,
         imu_mappings=imu_mappings,
         test_code=True,
         optimize_all=False)
 
+    # Data Keys
     pose_names = list(data.dynamic.keys())
     joint_names = list(data.dynamic[pose_names[0]].keys())
     imu_names = list(data.dynamic[pose_names[0]][joint_names[0]].keys())
     print(pose_names, joint_names, imu_names)
 
-    # methods = ['analytical', 'mittendorfer', 'normal_mittendorfer']
+    # Methods to Compare
     methods = ['analytical', 'mittendorfer']
-    # methods = ['mittendorfer']
 
     for i_su, su in enumerate(imu_names):
+        # joint which i_su th SU is attached to
         i_joint = kinematic_chain.su_joint_dict[i_su]
         for pose in pose_names:
+            # Consider 2 previous joints
             for rotate_joint in range(max(0, i_joint-2), i_joint+1):
                 joint = joint_names[rotate_joint]
 
+                # Break up the data
                 d = data.dynamic[pose][joint][su]
                 measured_As = d[:, :3]
                 joints = d[:, 3:10]
@@ -252,15 +317,17 @@ def verify_estimated_accelerations_for_dynamic_datacollection(args):
                 max_angular_velocity = d[0, 12]
                 joint_angular_velocities = d[:, 13]
 
+                # Prepare for plotting
                 y_dict = {'Measured': measured_As}
                 print(f'[{su}_{pose}_{joint}]')
 
                 n_data = d.shape[0]
                 for method in methods:
                     estimate_As = []
-                    additions = []
                     for i_data in range(n_data):
+                        # First Set current Pose (joints)
                         kinematic_chain.set_poses(joints[i_data], end_joint=i_joint)
+                        # Estimate current acceleration wrt the data
                         estimate_A = estimate_acceleration(
                             kinematic_chain=kinematic_chain,
                             i_rotate_joint=rotate_joint,
@@ -271,13 +338,12 @@ def verify_estimated_accelerations_for_dynamic_datacollection(args):
                             current_time=time[i_data],
                             method=method)
                         estimate_As.append(estimate_A)
-                        additions.append([joint_angular_velocities[i_data], joint_angular_accelerations[i_data]])
                     estimate_As = np.array(estimate_As)
+                    # Store to a dictionary
                     y_dict[method] = estimate_As
 
-                additions = np.array(additions)
-
-                plot_methods_at_once(
+                # Plot all methods in a same graph
+                plot_in_one_graph(
                     y_dict=y_dict,
                     ylabels=['ax', 'ay', 'az'],
                     xlabel='Time [s]',
@@ -286,8 +352,7 @@ def verify_estimated_accelerations_for_dynamic_datacollection(args):
                     show=True,
                     save=False)
 
-                # measured_As = np.hstack((measured_As, additions))
-                # estimate_As = np.hstack((estimate_As, additions))
+                # Plot 2 Comparative Methods side by side
                 # plot_side_by_side(
                 #     y1=measured_As,
                 #     y2=estimate_As,
